requirements.txt
jedi-language-server
python-telegram-bot
redis
black
pytest
pytest-mock
